Moving sub components to Profile component
