+++ components/profile/ProfileAchievements.tsx
@@ -3,7 +3,7 @@
 import React, { ReactElement } from 'react';
 import { Achievement } from '@/types/authorProfile';
 import { Tooltip } from '@/components/ui/Tooltip';
-import { getAchievementDetails, TIER_COLORS, TIER_INDICES } from './Achievements.utils';
+import { getAchievementDetails, TIER_COLORS, TIER_INDICES } from './ProfileAchievements.utils';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { cn } from '@/utils/styles';
 import { faTrophyStar } from '@fortawesome/pro-light-svg-icons';
